more corrections from PR review... tests are failing
