+++ pkg/server/v3/server_test.go
@@ -30,7 +30,6 @@ import (
 	"github.com/envoyproxy/go-control-plane/pkg/cache/types"
 	"github.com/envoyproxy/go-control-plane/pkg/cache/v3"
 	rsrc "github.com/envoyproxy/go-control-plane/pkg/resource/v3"
-	"github.com/envoyproxy/go-control-plane/pkg/server/stream/v3"
 	"github.com/envoyproxy/go-control-plane/pkg/server/v3"
 	"github.com/envoyproxy/go-control-plane/pkg/test/resource/v3"
 )
@@ -63,54 +62,6 @@ func (config *mockConfigWatcher) CreateWatch(req *discovery.DiscoveryRequest) (c
 			config.watches -= 1
 		}
 	}
-	return out, nil
-}
-
-func (config *mockConfigWatcher) CreateDeltaWatch(req *discovery.DeltaDiscoveryRequest, st *stream.StreamState) (chan cache.DeltaResponse, func()) {
-	config.deltaCounts[req.TypeUrl] = config.deltaCounts[req.TypeUrl] + 1
-
-	// Create our out watch channel to return with a buffer of one
-	out := make(chan cache.DeltaResponse, 1)
-
-	if len(config.deltaResponses[req.TypeUrl]) > 0 {
-		res := config.deltaResponses[req.TypeUrl][0]
-		var subscribed []types.Resource
-
-		// Only return back the subscribed resources to our request type
-		r, _ := res.GetDeltaDiscoveryResponse()
-		if len(req.GetResourceNamesSubscribe()) != 0 {
-			for _, resource := range r.Resources {
-				for _, alias := range req.GetResourceNamesSubscribe() {
-					if resource.GetName() == alias {
-						subscribed = append(subscribed, resource)
-					}
-				}
-			}
-		} else {
-			// We do this to handle the wildcard situation (just return all for testing)
-			for _, resource := range r.Resources {
-				subscribed = append(subscribed, resource)
-			}
-		}
-
-		// We should only send back subscribed resources here
-		out <- &cache.RawDeltaResponse{
-			DeltaRequest:      req,
-			Resources:         subscribed,
-			SystemVersionInfo: "",
-			NextVersionMap:    st.ResourceVersions,
-		}
-
-	} else if config.closeWatch {
-		close(out)
-	} else {
-		config.deltaWatches += 1
-		return out, func() {
-			close(out)
-			config.deltaWatches -= 1
-		}
-	}
-
 	return out, nil
 }
 
@@ -189,65 +140,6 @@ func makeMockStream(t *testing.T) *mockStream {
 		ctx:  context.Background(),
 		sent: make(chan *discovery.DiscoveryResponse, 10),
 		recv: make(chan *discovery.DiscoveryRequest, 10),
-	}
-}
-
-type mockDeltaStream struct {
-	t         *testing.T
-	ctx       context.Context
-	recv      chan *discovery.DeltaDiscoveryRequest
-	sent      chan *discovery.DeltaDiscoveryResponse
-	nonce     int
-	sendError bool
-	grpc.ServerStream
-}
-
-func (stream *mockDeltaStream) Context() context.Context {
-	return stream.ctx
-}
-
-func (stream *mockDeltaStream) Send(resp *discovery.DeltaDiscoveryResponse) error {
-	// check that nonce is monotonically incrementing
-	stream.nonce = stream.nonce + 1
-	if resp.Nonce != fmt.Sprintf("%d", stream.nonce) {
-		stream.t.Errorf("Nonce => got %q, want %d", resp.Nonce, stream.nonce)
-	}
-	// check resources are non-empty
-	if len(resp.Resources) == 0 {
-		stream.t.Error("Resources => got none, want non-empty")
-	}
-	// check that type URL matches in resources
-	if resp.TypeUrl == "" {
-		stream.t.Error("TypeUrl => got none, want non-empty")
-	}
-
-	for _, res := range resp.Resources {
-		if res.Resource.TypeUrl != resp.TypeUrl {
-			stream.t.Errorf("TypeUrl => got %q, want %q", res.Resource.TypeUrl, resp.TypeUrl)
-		}
-	}
-
-	stream.sent <- resp
-	if stream.sendError {
-		return errors.New("send error")
-	}
-	return nil
-}
-
-func (stream *mockDeltaStream) Recv() (*discovery.DeltaDiscoveryRequest, error) {
-	req, more := <-stream.recv
-	if !more {
-		return nil, errors.New("empty")
-	}
-	return req, nil
-}
-
-func makeMockDeltaStream(t *testing.T) *mockDeltaStream {
-	return &mockDeltaStream{
-		t:    t,
-		ctx:  context.Background(),
-		sent: make(chan *discovery.DeltaDiscoveryResponse, 10),
-		recv: make(chan *discovery.DeltaDiscoveryRequest, 10),
 	}
 }
 
@@ -333,61 +225,6 @@ func makeResponses() map[string][]cache.Response {
 				Version:   "7",
 				Resources: []types.ResourceWithTtl{{Resource: opaque}},
 				Request:   &discovery.DiscoveryRequest{TypeUrl: opaqueType},
-			},
-		},
-	}
-}
-
-func makeDeltaResponses() map[string][]cache.DeltaResponse {
-	return map[string][]cache.DeltaResponse{
-		rsrc.EndpointType: {
-			&cache.RawDeltaResponse{
-				Resources:         []types.Resource{endpoint},
-				DeltaRequest:      &discovery.DeltaDiscoveryRequest{TypeUrl: rsrc.EndpointType},
-				SystemVersionInfo: "1",
-			},
-		},
-		rsrc.ClusterType: {
-			&cache.RawDeltaResponse{
-				Resources:         []types.Resource{cluster},
-				DeltaRequest:      &discovery.DeltaDiscoveryRequest{TypeUrl: rsrc.ClusterType},
-				SystemVersionInfo: "2",
-			},
-		},
-		rsrc.RouteType: {
-			&cache.RawDeltaResponse{
-				Resources:         []types.Resource{route},
-				DeltaRequest:      &discovery.DeltaDiscoveryRequest{TypeUrl: rsrc.RouteType},
-				SystemVersionInfo: "3",
-			},
-		},
-		rsrc.ListenerType: {
-			&cache.RawDeltaResponse{
-				Resources:         []types.Resource{listener},
-				DeltaRequest:      &discovery.DeltaDiscoveryRequest{TypeUrl: rsrc.ListenerType},
-				SystemVersionInfo: "4",
-			},
-		},
-		rsrc.SecretType: {
-			&cache.RawDeltaResponse{
-				SystemVersionInfo: "5",
-				Resources:         []types.Resource{secret},
-				DeltaRequest:      &discovery.DeltaDiscoveryRequest{TypeUrl: rsrc.SecretType},
-			},
-		},
-		rsrc.RuntimeType: {
-			&cache.RawDeltaResponse{
-				SystemVersionInfo: "6",
-				Resources:         []types.Resource{runtime},
-				DeltaRequest:      &discovery.DeltaDiscoveryRequest{TypeUrl: rsrc.RuntimeType},
-			},
-		},
-		// Pass-through type (xDS does not exist for this type)
-		opaqueType: {
-			&cache.RawDeltaResponse{
-				SystemVersionInfo: "7",
-				Resources:         []types.Resource{opaque},
-				DeltaRequest:      &discovery.DeltaDiscoveryRequest{TypeUrl: opaqueType},
 			},
 		},
 	}
@@ -481,58 +318,6 @@ func TestResponseHandlers(t *testing.T) {
 				close(resp.recv)
 				if want := map[string]int{typ: 1}; !reflect.DeepEqual(want, config.counts) {
 					t.Errorf("watch counts => got %v, want %v", config.counts, want)
-				}
-			case <-time.After(1 * time.Second):
-				t.Fatalf("got no response")
-			}
-		})
-	}
-}
-
-func TestDeltaResponseHandlers(t *testing.T) {
-	for _, typ := range testTypes {
-		t.Run(typ, func(t *testing.T) {
-			config := makeMockConfigWatcher()
-			config.deltaResponses = makeDeltaResponses()
-			s := server.NewServer(context.Background(), config, server.CallbackFuncs{})
-
-			resp := makeMockDeltaStream(t)
-			// This should put through a wildcard request since we aren't subscribing to anything
-			resp.recv <- &discovery.DeltaDiscoveryRequest{Node: node, TypeUrl: typ, ResourceNamesSubscribe: []string{}}
-
-			go func() {
-				var err error
-				switch typ {
-				case rsrc.EndpointType:
-					err = s.DeltaEndpoints(resp)
-				case rsrc.ClusterType:
-					err = s.DeltaClusters(resp)
-				case rsrc.RouteType:
-					err = s.DeltaRoutes(resp)
-				case rsrc.ListenerType:
-					err = s.DeltaListeners(resp)
-				case rsrc.SecretType:
-					err = s.DeltaSecrets(resp)
-				case rsrc.RuntimeType:
-					err = s.DeltaRuntime(resp)
-				case opaqueType:
-					err = s.DeltaAggregatedResources(resp)
-				}
-
-				if err != nil {
-					t.Errorf("Delta() => got \"%v\", want no error", err)
-				}
-			}()
-
-			select {
-			case res := <-resp.sent:
-				close(resp.recv)
-				if want := map[string]int{typ: 1}; !reflect.DeepEqual(want, config.deltaCounts) {
-					t.Errorf("watch counts => got %v, want %v", config.deltaCounts, want)
-				}
-
-				if v := res.GetSystemVersionInfo(); v != "" {
-					t.Errorf("should've had an emtpy version for first request, got %s", v)
 				}
 			case <-time.After(1 * time.Second):
 				t.Fatalf("got no response")
@@ -670,29 +455,6 @@ func TestWatchClosed(t *testing.T) {
 	}
 }
 
-func TestDeltaWatchClosed(t *testing.T) {
-	for _, typ := range testTypes {
-		t.Run(typ, func(t *testing.T) {
-			config := makeMockConfigWatcher()
-			config.closeWatch = true
-			s := server.NewServer(context.Background(), config, server.CallbackFuncs{})
-
-			resp := makeMockDeltaStream(t)
-			resp.recv <- &discovery.DeltaDiscoveryRequest{
-				Node:    node,
-				TypeUrl: typ,
-			}
-
-			// check that response fails since watch gets closed
-			if err := s.DeltaAggregatedResources(resp); err == nil {
-				t.Error("DeltaAggregatedResources() => got no error, want watch failed")
-			}
-
-			close(resp.recv)
-		})
-	}
-}
-
 func TestSendError(t *testing.T) {
 	for _, typ := range testTypes {
 		t.Run(typ, func(t *testing.T) {
@@ -711,31 +473,6 @@ func TestSendError(t *testing.T) {
 			// check that response fails since send returns error
 			if err := s.StreamAggregatedResources(resp); err == nil {
 				t.Error("Stream() => got no error, want send error")
-			}
-
-			close(resp.recv)
-		})
-	}
-}
-
-func TestSendDeltaError(t *testing.T) {
-	for _, typ := range testTypes {
-		t.Run(typ, func(t *testing.T) {
-			config := makeMockConfigWatcher()
-			config.deltaResponses = makeDeltaResponses()
-			s := server.NewServer(context.Background(), config, server.CallbackFuncs{})
-
-			// make a request
-			resp := makeMockDeltaStream(t)
-			resp.sendError = true
-			resp.recv <- &discovery.DeltaDiscoveryRequest{
-				Node:    node,
-				TypeUrl: typ,
-			}
-
-			// check that response fails since send returns error
-			if err := s.DeltaAggregatedResources(resp); err == nil {
-				t.Error("DeltaAggregatedResources() => got no error, want send error")
 			}
 
 			close(resp.recv)
@@ -833,62 +570,6 @@ func TestAggregatedHandlers(t *testing.T) {
 					rsrc.ListenerType: 1,
 				}; !reflect.DeepEqual(want, config.counts) {
 					t.Errorf("watch counts => got %v, want %v", config.counts, want)
-				}
-
-				// got all messages
-				return
-			}
-		case <-time.After(1 * time.Second):
-			t.Fatalf("got %d messages on the stream, not 4", count)
-		}
-	}
-}
-
-func TestDeltaAggregatedHandlers(t *testing.T) {
-	config := makeMockConfigWatcher()
-	config.deltaResponses = makeDeltaResponses()
-	resp := makeMockDeltaStream(t)
-
-	resp.recv <- &discovery.DeltaDiscoveryRequest{
-		Node:    node,
-		TypeUrl: rsrc.ListenerType,
-	}
-	// Delta compress node
-	resp.recv <- &discovery.DeltaDiscoveryRequest{
-		Node:    node,
-		TypeUrl: rsrc.ClusterType,
-	}
-	resp.recv <- &discovery.DeltaDiscoveryRequest{
-		Node:                   node,
-		TypeUrl:                rsrc.EndpointType,
-		ResourceNamesSubscribe: []string{clusterName},
-	}
-	resp.recv <- &discovery.DeltaDiscoveryRequest{
-		TypeUrl:                rsrc.RouteType,
-		ResourceNamesSubscribe: []string{routeName},
-	}
-
-	s := server.NewServer(context.Background(), config, server.CallbackFuncs{})
-	go func() {
-		if err := s.DeltaAggregatedResources(resp); err != nil {
-			t.Errorf("DeltaAggregatedResources() => got %v, want no error", err)
-		}
-	}()
-
-	count := 0
-	for {
-		select {
-		case <-resp.sent:
-			count++
-			if count >= 4 {
-				close(resp.recv)
-				if want := map[string]int{
-					rsrc.EndpointType: 1,
-					rsrc.ClusterType:  1,
-					rsrc.RouteType:    1,
-					rsrc.ListenerType: 1,
-				}; !reflect.DeepEqual(want, config.deltaCounts) {
-					t.Errorf("watch counts => got %v, want %v", config.deltaCounts, want)
 				}
 
 				// got all messages
